renewed navigator base interface
