+++ nav2_bt_navigator/include/nav2_bt_navigator/navigators/navigate_through_poses.hpp
@@ -21,7 +21,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_action/rclcpp_action.hpp"
 #include "geometry_msgs/msg/pose_stamped.hpp"
-#include "nav2_core/navigator.hpp"
+#include "nav2_bt_navigator/navigator.hpp"
 #include "nav2_msgs/action/navigate_through_poses.hpp"
 #include "nav_msgs/msg/path.hpp"
 #include "nav2_util/robot_utils.hpp"
@@ -36,7 +36,7 @@ namespace nav2_bt_navigator
  * @brief A navigator for navigating to a a bunch of intermediary poses
  */
 class NavigateThroughPosesNavigator
-  : public nav2_core::Navigator<nav2_msgs::action::NavigateThroughPoses>
+  : public nav2_bt_navigator::Navigator<nav2_msgs::action::NavigateThroughPoses>
 {
 public:
   using ActionT = nav2_msgs::action::NavigateThroughPoses;
@@ -48,14 +48,12 @@ public:
   NavigateThroughPosesNavigator()
   : Navigator() {}
 
+protected:
   /**
    * @brief A configure state transition to configure navigator's state
    * @param node Weakptr to the lifecycle node
-   * @param odom_smoother Object to get current smoothed robot's speed
    */
-  bool configure(
-    rclcpp_lifecycle::LifecycleNode::WeakPtr node,
-    std::shared_ptr<nav2_util::OdomSmoother> odom_smoother) override;
+  bool onConfigure(rclcpp_lifecycle::LifecycleNode::WeakPtr node) override;
 
   /**
    * @brief Get action name for this navigator
@@ -70,7 +68,6 @@ public:
    */
   std::string getDefaultBTFilepath(rclcpp_lifecycle::LifecycleNode::WeakPtr node) override;
 
-protected:
   /**
    * @brief A callback to be called when a new goal is received by the BT action server
    * Can be used to check if goal is valid and put values on
@@ -106,6 +103,8 @@ protected:
    * @brief Goal pose initialization on the blackboard
    */
   void initializeGoalPoses(ActionT::Goal::ConstSharedPtr goal);
+
+  
 
   rclcpp::Time start_time_;
   std::string goals_blackboard_id_;
