+++ nav2_bt_navigator/src/navigators/navigate_to_pose.cpp
@@ -23,9 +23,7 @@ namespace nav2_bt_navigator
 {
 
 bool
-NavigateToPoseNavigator::configure(
-  rclcpp_lifecycle::LifecycleNode::WeakPtr parent_node,
-  std::shared_ptr<nav2_util::OdomSmoother> odom_smoother)
+NavigateToPoseNavigator::onConfigure(rclcpp_lifecycle::LifecycleNode::WeakPtr parent_node)
 {
   start_time_ = rclcpp::Time(0);
   auto node = parent_node.lock();
@@ -41,9 +39,6 @@ NavigateToPoseNavigator::configure(
   }
 
   path_blackboard_id_ = node->get_parameter("path_blackboard_id").as_string();
-
-  // Odometry smoother object for getting current speed
-  odom_smoother_ = odom_smoother;
 
   self_client_ = rclcpp_action::create_client<ActionT>(node, getName());
